+++ src/vulkanCommands.h
@@ -17,7 +17,7 @@
 class VulkanCommands
 {
 public: // Members:
-	size_t size;	// should be equal to framesInFlight
+	uint32_t size;	// should be equal to framesInFlight
 	std::vector<VkCommandPool> pools;
 	std::vector<VkCommandBuffer> buffers;
 
@@ -25,7 +25,7 @@ private: // Members:
 	VulkanLogicalDevice* logicalDevice;
 
 public: // Methods:
-	VulkanCommands(size_t size, VulkanLogicalDevice* logicalDevice, VulkanQueue queue);
+	VulkanCommands(uint32_t size, VulkanLogicalDevice* logicalDevice, VulkanQueue queue);
 	~VulkanCommands();
 };
